Separate poi controllers
